+++ docs/ReleaseNotes.md
@@ -1,6 +1,7 @@
 // Before any commit, use the command line: npx tsc --noEmit
 
-# [1,2] File Explore
+# [1,2] Camera Picture Preview
 
-- Some debugging colors was removed (Forgot to remove before last commit)
-- Horizontal Camera UI slightly Ajusted
+- Added Zoom to preview
+- Preview now can be dragged
+- Enable <PictureInput /> to use the preview layer to already saved pictures